+++ src/Development/IDE/Types/Options.hs
@@ -15,6 +15,8 @@ module Development.IDE.Types.Options
   ) where
 
 import Data.Maybe
+import Development.Shake
+import Development.IDE.GHC.Util
 import           GHC hiding (parseModule, typecheckModule)
 import           GhcPlugins                     as GHC hiding (fst3, (<>))
 import qualified Language.Haskell.LSP.Types.Capabilities as LSP
@@ -23,6 +25,10 @@ data IdeOptions = IdeOptions
   { optPreprocessor :: GHC.ParsedSource -> IdePreprocessedSource
     -- ^ Preprocessor to run over all parsed source trees, generating a list of warnings
     --   and a list of errors, along with a new parse tree.
+  , optGhcSession :: Action (FilePath -> Action HscEnvEq)
+    -- ^ Setup a GHC session for a given file, e.g. @Foo.hs@.
+    --   For the same 'ComponentOptions' from hie-bios, the resulting function will be applied once per file.
+    --   It is desirable that many files get the same 'HscEnvEq', so that more IDE features work.
   , optPkgLocationOpts :: IdePkgLocationOptions
     -- ^ How to locate source and @.hie@ files given a module name.
   , optExtensions :: [String]
@@ -65,9 +71,10 @@ clientSupportsProgress :: LSP.ClientCapabilities -> IdeReportProgress
 clientSupportsProgress caps = IdeReportProgress $ fromMaybe False $
     LSP._workDoneProgress =<< LSP._window (caps :: LSP.ClientCapabilities)
 
-defaultIdeOptions :: IdeOptions
-defaultIdeOptions = IdeOptions
+defaultIdeOptions :: Action (FilePath -> Action HscEnvEq) -> IdeOptions
+defaultIdeOptions session = IdeOptions
     {optPreprocessor = IdePreprocessedSource [] []
+    ,optGhcSession = session
     ,optExtensions = ["hs", "lhs"]
     ,optPkgLocationOpts = defaultIdePkgLocationOptions
     ,optThreads = 0
